requirements.txt
streamlit==1.45.0
tensorflow==2.19.0
numpy==2.1.3
pillow==11.2.1
opencv-python
matplotlib==3.10.1
scikit-learn==1.6.1
seaborn==0.13.2
